web-instruccion-login-google/codigo/requirements.txt
Flask
Flask-Login
oauthlib
pyOpenSSL
python-dotenv
requests